Implement User roles

See merge request clearview/root!326

+++ api/src/services/facade/SpaceFacade.ts
@@ -5,6 +5,8 @@ import { SpaceService, UserSpaceService, NodeService, UserService } from '../'
 import { ServiceFactory } from '../factory/ServiceFactory'
 import { clientError, HttpErrName, HttpStatusCode } from '../../response/errors'
 import { Node } from '../../database/entities/Node'
+import { SpaceUserRole } from '../../types/spaceUser'
+import { SpaceUserUpdateValue } from '../../values/spaceUser'
 
 export class SpaceFacade {
   private spaceService: SpaceService
@@ -42,7 +44,7 @@ export class SpaceFacade {
   async createSpace(data: SpaceCreateValue): Promise<Space> {
     const space = await this.spaceService.create(data)
 
-    await this.userSpaceService.add(space.userId, space.id)
+    await this.userSpaceService.add(space.userId, space.id, SpaceUserRole.Admin)
 
     await this.nodeService.createSpaceRootNode(space.id, space.userId)
     await this.nodeService.createSpaceArchiveNode(space.id)
@@ -54,7 +56,7 @@ export class SpaceFacade {
     return this.spaceService.update(data, spaceId)
   }
 
-  async removeUserFromSpace(userId: number, spaceId: number): Promise<UserToSpace> {
+  async updateSpaceUser(data: SpaceUserUpdateValue, userId: number, spaceId: number): Promise<UserToSpace> {
     const user = await this.userService.requireUserById(userId)
     const space = await this.spaceService.requireSpaceById(spaceId)
 
@@ -63,11 +65,26 @@ export class SpaceFacade {
     }
 
     if (user.id === space.userId) {
-      throw clientError('Can not remove space owner from space', HttpErrName.InvalidRequest, HttpStatusCode.NotAllowed)
+      throw clientError('Can not update space owner role', HttpErrName.NotAllowed, HttpStatusCode.Forbidden)
     }
 
-    const userSpace = this.userSpaceService.remove(userId, spaceId)
+    const result = await this.userSpaceService.update(data, userId, spaceId)
+    return result
+  }
 
+  async removeSpaceUser(userId: number, spaceId: number): Promise<UserToSpace> {
+    const user = await this.userService.requireUserById(userId)
+    const space = await this.spaceService.requireSpaceById(spaceId)
+
+    if ((await this.userSpaceService.isUserInSpace(userId, spaceId)) === false) {
+      throw clientError('User not found in space', HttpErrName.EntityNotFound, HttpStatusCode.NotFound)
+    }
+
+    if (user.id === space.userId) {
+      throw clientError('Can not remove space owner from space', HttpErrName.InvalidRequest, HttpStatusCode.Forbidden)
+    }
+
+    const userSpace = await this.userSpaceService.remove(userId, spaceId)
     return userSpace
   }
 }
